requirements.txt
#Web framework
fastapi
uvicorn[standard]

#Database 
sqlalchemy
pydantic
bcrypt


#ML & Data Science
scikit-learn
pandas
numpy
joblib
matplotlib
seaborn

# Time Series Forecasting 
prophet
statsmodels

# Others 
python-dotenv

